Vendor `fast-xml-parser`, minor refactor of vendoring
