+++ .eslintignore
@@ -1,4 +1,4 @@
 scratch
 test/mock/
 bench/tmp/
-**/_vendor.*js
+**/_vendor/*
